requirements.txt
requests
google-cloud-bigquery
dbt-bigquery
pytest